third_party/WebKit/LayoutTests/csspaint/style-first-letter-pseudo-expected.txt

CONSOLE MESSAGE: line 8: color: rgb(255, 0, 0)
CONSOLE MESSAGE: line 8: line-height: 2px
Foo
This tests the style information in the paint callback.

See the devtools console for test output. The console should log:

color: rgb(255, 0, 0)
line-height: 2px
